feat: Strategy Pattern을 사용한 로그인 구현 - AuthenticationStrategy protocol 정의
