+++ yeogijeogi/auth/AuthenticationStrategy.swift
@@ -1,0 +1,6 @@
+import FirebaseAuth
+
+protocol AuthenticationStrategy {
+    func signIn(completion: @escaping (Result<AuthCredential, Error>) -> Void)
+    func signOut()
+}
