--- yeogijeogi/utils/Authenticator.swift
@@ -1,61 +1,0 @@
-import Firebase
-import FirebaseAuth
-import GoogleSignIn
-
-class Authenticator: ObservableObject {
-    @Published var signState: SignState = .signOut
-    
-    func checkSignState() {
-        signState = Auth.auth().currentUser != nil ? .signIn : .signOut
-    }
-    
-    func signIn() {
-        if GIDSignIn.sharedInstance.hasPreviousSignIn() {
-            GIDSignIn.sharedInstance.restorePreviousSignIn { [unowned self] user, error in
-                self.authenticateUser(for: user, with: error)
-            }
-        } else {
-            guard let clientID = FirebaseApp.app()?.options.clientID else { return }
-            
-            let configuration = GIDConfiguration(clientID: clientID)
-            GIDSignIn.sharedInstance.configuration = configuration
-            
-            guard let windowScene = UIApplication.shared.connectedScenes.first as? UIWindowScene else { return }
-            guard let rootViewController = windowScene.windows.first?.rootViewController else { return }
-            
-            GIDSignIn.sharedInstance.signIn(withPresenting: rootViewController) { [unowned self] result, error in
-                guard let result = result else { return }
-                self.authenticateUser(for: result.user, with: error)
-            }
-        }
-    }
-    
-    private func authenticateUser(for user: GIDGoogleUser?, with error: Error?) {
-        if let error = error {
-            print(error.localizedDescription)
-            return
-        }
-        
-        guard let accessToken = user?.accessToken.tokenString, let idToken = user?.idToken?.tokenString else { return }
-        let credential = GoogleAuthProvider.credential(withIDToken: idToken, accessToken: accessToken)
-        
-        Auth.auth().signIn(with: credential) { _, error in
-            if let error = error {
-                print(error.localizedDescription)
-            } else {
-                self.signState = .signIn
-            }
-        }
-    }
-    
-    func signOut() {
-        GIDSignIn.sharedInstance.signOut()
-        
-        do {
-            try Auth.auth().signOut()
-            signState = .signOut
-        } catch {
-            print(error.localizedDescription)
-        }
-    }
-}
